WIP reduction in eth_call by moving more direct contract calls to the snapshot
Also re-arranged some test code to avoid importing modules with tests. Around 19 tests failing atm.
